Rename packages to be more PEP8-ble
