+++ test/locators/checkboxes.py
