+++ test/locators/icons.py
